Move getTenant method to util class
