+++ src/main/java/org/folio/inventory/dataimport/handlers/actions/MarcBibMatchedPostProcessingEventHandler.java
@@ -30,7 +30,7 @@ import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.folio.ActionProfile.FolioRecord.HOLDINGS;
 import static org.folio.ActionProfile.FolioRecord.INSTANCE;
 import static org.folio.DataImportEventTypes.DI_SRS_MARC_BIB_RECORD_MATCHED_READY_FOR_POST_PROCESSING;
-import static org.folio.inventory.dataimport.handlers.matching.util.EventHandlingUtil.constructContext;
+import static org.folio.inventory.dataimport.handlers.matching.util.EventHandlingUtil.getTenant;
 import static org.folio.inventory.dataimport.util.LoggerUtil.logParametersEventHandler;
 
 public class MarcBibMatchedPostProcessingEventHandler implements EventHandler {
@@ -43,7 +43,6 @@ public class MarcBibMatchedPostProcessingEventHandler implements EventHandler {
   private static final String ERROR_HOLDING_MSG = "Error loading inventory holdings for MARC BIB";
   private static final String MATCHED_MARC_BIB_KEY = "MATCHED_MARC_BIBLIOGRAPHIC";
   private final Storage storage;
-  private static final String CENTRAL_TENANT_ID = "CENTRAL_TENANT_ID";
 
   public MarcBibMatchedPostProcessingEventHandler(Storage storage) {
     this.storage = storage;
@@ -108,14 +107,6 @@ public class MarcBibMatchedPostProcessingEventHandler implements EventHandler {
       future.completeExceptionally(e);
     }
     return future;
-  }
-
-  public static String getTenant(DataImportEventPayload payload) {
-    String centralTenantId = payload.getContext().get(CENTRAL_TENANT_ID);
-    if (centralTenantId != null) {
-      return centralTenantId;
-    }
-    return payload.getTenant();
   }
 
   @Override
